[LANG] Expand tl.target_info and fail gracefully on CPU
This moves some target_info functions from kernels into the language.
I also update `current_target()` to return `None` when there is no
supported device found, which allows the target function to be called
e.g. at import time.
